requirements.txt
llama-index==0.12.11
openai==1.59.8
langchain==0.3.14
requests==2.32.3
rdflib==7.1.3
nltk==3.9.1
Flask==3.1.0
python-dotenv==1.0.1
numpy<2
pandas~=2.2.3
matplotlib~=3.10.0
plotly~=5.24.1
scipy~=1.15.1
Werkzeug~=3.1.3